Update 2.5 fix error add item to cart, get product by store

+++ drinking-be-v2/Models/Store.cs
@@ -26,7 +26,7 @@ public partial class Store : ISoftDelete
 
     public StoreStatusEnum Status { get; set; } = StoreStatusEnum.ComingSoon;
 
-    public byte? SortOrder { get; set; }
+    public short? SortOrder { get; set; }
 
     public bool MapVerified { get; set; } = false;
